requirements.txt
django==4.2
requests==2.32.2
djangorestframework==3.15.2
gunicorn==23.0.0
